withinr: split the bridge into api/config/convert/results modules
Pure code motion: mirror the module layout of crates/within-py so a
change to the Python glue points at the same-named file in the R
glue. lib.rs is registration-only, composed from per-module
extendr_module! blocks; api holds the solve entry points and the
solver handle, config the R->native config parsing and the
preconditioner handle, convert the shared coercion helpers, results
the native->list conversions.
